fix: 32bits build issues
it also closes issue #3.

+++ pkg/config/bytes_test.go
@@ -36,7 +36,7 @@ func Test_formatBytes(t *testing.T) {
 	}
 	tests := []struct {
 		name string
-		args int
+		args Byte
 		want string
 	}{
 		{"none", -1, "-1"},
